requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.6.0
ansicon==1.89.0
anthropic==0.25.2
anyio==4.3.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
autobahn==24.4.2
Automat==25.4.16
awscli==1.40.18
awsebcli==3.24.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.1
blessed==1.21.0
blinker==1.7.0
boto3==1.38.19
botocore==1.38.20
cachetools==5.3.3
celery==5.5.2
cement==2.10.14
certifi==2024.2.2
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
clr_loader==0.2.7.post0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==42.0.5
cssselect2==0.7.0
curl_cffi==0.6.2
daphne==4.2.0
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.0.2
django-celery-beat==2.8.0
django-cors-headers==4.3.1
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dnspython==2.6.1
docutils==0.19
duckduckgo_search==5.3.0
eventlet==0.36.1
fabric==3.2.2
fastlogging==1.0.0
ffmpeg==1.4
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.1
fsspec==2024.3.1
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.1
googleapis-common-protos==1.63.0
gotrue==2.12.0
greenlet==3.0.3
groq==0.5.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
iniconfig==2.0.0
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jinxed==1.3.0
jmespath==1.0.1
joblib==1.4.0
keybert==0.8.4
kombu==5.5.3
llvmlite==0.44.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
mdurl==0.1.2
mistletoe==1.3.0
mistralai==0.0.8
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mysql-connector-python==9.3.0
mysqlclient==2.2.7
netlify-py==0.1.0
networkx==3.3
npm==0.1.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.8
openai==1.20.0
openai-whisper==20240930
opencv-python==4.11.0.86
optional-django==0.1.0
orjson==3.10.1
oscrypto==1.3.0
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pdfminer.six==20231228
pillow==10.3.0
platformdirs==4.3.6
playwright==1.43.0
pluggy==1.4.0
postgrest==1.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyee==11.1.0
Pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.1.2
pypdf==4.2.0
pypiwin32==223
pypng==0.20220715.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-mock==3.14.0
pytest-playwright==0.4.4
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-slugify==8.0.4
python-socketio==5.11.2
python-telegram-bot==21.11.1
python3-openid==3.2.0
pythonnet==3.0.5
pytz==2025.2
pywin32==310
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.2
realtime==2.4.2
redis==4.5.5
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.7.2
s3transfer==0.12.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
sentence-transformers==2.6.1
service-identity==24.2.0
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.5
SQLAlchemy==2.0.29
sqlmodel==0.0.16
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
svglib==1.5.1
sympy==1.12
tabulate==0.9.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.15
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
Werkzeug==3.0.2
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
xhtml2pdf==0.2.15
yarl==1.18.3
zope.event==5.0
zope.interface==6.3
